Add the ability to express holes in a trace.
Currently we stop mapping a trace at the first (non-prefix) unmappable
block. This is incorrect in the long run, as the trace may not be
complete. For example, it could be that there's a call to external code
for which we have no IR.

This change allows us to express holes in a trace, which we then leave
up to the trace compiler to interpret as it sees fit.

+++ c_tests/tests/mapper_funcs_branches.c
@@ -25,7 +25,7 @@ int main(int argc, char **argv) {
   void *tr = __yktrace_stop_tracing(tt);
 
   size_t len = __yktrace_irtrace_len(tr);
-  for (int i = 0; i < len; i++) {
+  for (size_t i = 0; i < len; i++) {
     char *func_name = NULL;
     size_t bb = 0;
     __yktrace_irtrace_get(tr, i, &func_name, &bb);
